+++ src/globalContext.js
@@ -6,9 +6,9 @@ function GlobalProvider({ children }) {
     const [ gameData, setGameData ] = useState(null)
     const [userData, setUserData] = useState({
         points: {
-          gross: 5000,
+          gross: 100000,
           spent: 0,
-          net: 0
+          net: 100000
         },
         bonusData: [
             {
@@ -20,7 +20,7 @@ function GlobalProvider({ children }) {
           },
           {
             level: 1,
-            reward: 100,
+            reward: 200,
           }
         ],
         time: 0
@@ -43,15 +43,41 @@ function GlobalProvider({ children }) {
 
     useEffect(() => {
 
-        if (userData.points.net !== (userData.points.gross - userData.points.spent)) {
+        if (gameData === null) return
+
+        const correctNet = userData.points.gross - userData.points.spent;
+        if (userData.points.net !== correctNet) {
           setUserData(data => {
             return {...data,
               points: {...data.points,
-                net: data.points.gross - data.points.spent
+               net: correctNet
               }
             }
           })
         }
+    
+        const correctRapidInputReward = Math.ceil(gameData.bonusData[2].recursion.base * (gameData.bonusData[2].recursion.growthRate ** userData.bonusData[2].level));
+        if (userData.bonusData[2].reward !== correctRapidInputReward) {
+            /* console.log('ran :',userData.bonusData[2].reward,correctRapidInputReward) */
+            const newBonusArray = userData.bonusData.map((bonus,index) => index === 2 ? {...bonus, reward: correctRapidInputReward} : bonus);
+            setUserData(data => {
+                return {...data,
+                    bonusData: newBonusArray
+                }
+            })
+        }
+
+        const correctLuckyLetter = gameData.bonusData[0].sequence[userData.bonusData[0].level];
+        if (userData.bonusData[0].letter !== correctLuckyLetter) {
+            /* console.log('ran :',userData.bonusData[2].reward,correctRapidInputReward) */
+            const newBonusArray = userData.bonusData.map((bonus,index) => index === 0 ? {...bonus, letter: correctLuckyLetter} : bonus);
+            setUserData(data => {
+                return {...data,
+                    bonusData: newBonusArray
+                }
+            })
+        }
+
       },[userData])
 
     return <globalContext.Provider value={{gameData: gameData, userData: userData, setUserData: setUserData}}>{children}</globalContext.Provider>